Move persistence files to own modules
