+++ snn_hfo_ieeg/stages/persistence/saving.py
@@ -1,0 +1,49 @@
+import os
+from pathlib import Path
+import scipy.io as sio
+from snn_hfo_ieeg.stages.persistence.utility import get_persistence_path
+
+
+def _is_namedtuple(obj) -> bool:
+    # Source: https://stackoverflow.com/a/62692640/5903309
+    return (
+        isinstance(obj, tuple) and
+        hasattr(obj, '_asdict') and
+        hasattr(obj, '_fields')
+    )
+
+
+def _is_list(obj) -> bool:
+    return (
+        hasattr(obj, '__iter__') and
+        not isinstance(obj, (str, dict))
+    )
+
+
+def _is_dict(obj) -> bool:
+    return isinstance(obj, dict)
+
+
+def _convert_to_dict(object):
+    if _is_namedtuple(object):
+        return _convert_to_dict(object._asdict())
+    if _is_list(object):
+        return [_convert_to_dict(_) for _ in object]
+    if _is_dict(object):
+        for key, item in object.items():
+            object[key] = _convert_to_dict(item)
+        return object
+    # Needed because of https://stackoverflow.com/questions/38934433/maximum-recursion-depth-exceeded-when-using-scipy-io-savemat
+    return str(object)
+
+
+def _create_parent_directory(path):
+    parent_directory = Path(path).parent.absolute()
+    os.makedirs(parent_directory, exist_ok=True)
+
+
+def save_hfo_detection(user_facing_hfo_detection, saving_path, metadata):
+    filepath = get_persistence_path(saving_path, metadata)
+    _create_parent_directory(filepath)
+    dictionary = _convert_to_dict(user_facing_hfo_detection)
+    sio.savemat(filepath, dictionary)
